Initial commit
This project is for a bit-serial CPU, written in VHDL, with a simulator
for the system written in C. It should eventually be targeted to run on
an actual FPGA after the simulation and design has been done.

This is the initial commit, it does not contain a working simulation (C
or VHDL) and the design *will* change as the current design is
inadequate. At the moment the repository is more a play ground for
ideas.