Added primitive assembler, program counter working
VHDL:
* The test bench is very minimal at the moment, but the framework is in
place. The program memory will need to be loaded from disk into Block
RAM and then squeezed out through a shift register soon, but that is a
task for later.
* The makefile has been improved, new targets and things.
* The program counter seems to be working.
* I wonder if this is the simplest way of designing this processor,
there seems to be an awful lot of repetition, and complication. It may
need a redesign once it hits the FPGA toolchain...

C/Simulator:
* A simple assembler has been added. It is lacking forward references
and many other amenities at the moment. But it has made the tiny test
program a lot more readable.