Made each cycle fixed length
Variable length cycles will make the external interface more difficult
to implement, so everything has been made to be fixed length.

As a note, perhaps functions should be used to implement the shifting so
it is easier to keep track of whether we are shifting things in the same
direction or not.