Mostly working VHDL simulation CPU
The CPU mostly works now, not all of the instructions have been tested,
and the 'less' instruction needs to be implemented.

Other changes:
* Added a memory module skeleton, this will eventually contain a Block
RAM module and Input/Output decoder, at the moment it only contains the
shift registers that interface between that module and the outside
world.
* A test plan has been added to the 'bit.asm' file.
* Test bench now includes the memory module, as does the makefile.
* The bit order that the CPU does operations in has now been fixed, it
does all operations where possible with the Least Significant Bit first.