Rearranged instruction set (again)
A top level VHDL file now contains the project, and a block RAM module,
which has been implemented, even if it is not correct at the moment.

The instruction set has been changed, yet again, it is a little more
compact and useful now. A few bugs have been fixed, there is now an
instruction for manipulating the CPU status flags (which can also be
used to reset or halt the processor). The less than instruction has been
replaced with a subtract instruction which effects the CPU status flags.

The CPU simulation is *almost* working correctly. There is now going to
be a lot of debugging work to do.

Some more TODO comments have also been added.