Added VHDL build system, CPU still not working
The CPU design is still not working correctly, however there have been a
number of improvements and it appears closer to working. There should
not have to be any radical changes to get the system working.

The design synthesises and is quite small, using about 1% of the Spartan
6 FPGA on the Nexys-3 board that the build currently targets. It of
course uses a block RAM as well, but the CPU itself is tiny. The first
task it to get it working, I will try to make it even smaller after.
There will no doubt be problems in synthesis after the simulation is
working as well.