Simulated CPU working!
The simulated version of the CPU is working, however this will not
synthesize into anything sensible. That will have to be fixed. The test
program does not exercise all instructions either, so there is still a
few things to test before it can be declared correct. However, this is a
good start.

The test program adds a 1 to a memory location (which starts as 3)
until it exceeds 15, then halts; it exercises addition, and, jumping,
loading, storing and (some) flag setting.

A few new directives and pseudo instructions have been added to the
assembler as well, including; 'nop', 'clr', a single argument 'invert',
allowing numbers to be directly embedded into the code, and 'set'.

The arithmetic in the simulator (for 'add' and 'sub' instructions) has
been corrected, but not tested, this adds in the previous carry (for
addition) or borrow flag (for subtraction) into the calculation.