Design synthesizes (untested)
The design synthesizes, hopefully into something sensible. It has not
actually been tested on an FPGA, and only the simulation seems to
work/has been tested. There is more work to do to get this into a
working state (specifically the I/O needs working on). The test program
is the only thing that has been run in the simulation still.

The change in this commit was to make the ram write occur on the falling
edge of the 'ie' enable.