UART code, made cpu smaller, delay, and more
The UART code has been added to the project, and LED output pins added
to the top level module (which may be removed...). The UART code has not
be integrated yet, but it would be a nice small test that the CPU is
working and a step closer to implementing a more useful system on top of
this CPU.

'after delay' has been added to the BCPU module, but not elsewhere in
the project.

The processor has been simplified a little, code moved around, but it is
functionally equivalent.

The processor has also been made to be smaller. The current utilization
hierarchy is:

+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Module             | Partition | Slices*       | Slice Reg     | LUTs          | LUTRAM        | BRAM/FIFO | DSP48A1 | BUFG  | BUFIO | BUFR  | DCM   | PLL_ADV   | Full Hierarchical  |
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| top/               |           | 0/38          | 0/113         | 0/105         | 0/4           | 0/4       | 0/0     | 1/1   | 0/0   | 0/0   | 0/0   | 0/0       | top                |
| +cpu               |           | 24/24         | 65/65         | 76/76         | 3/3           | 0/0       | 0/0     | 0/0   | 0/0   | 0/0   | 0/0   | 0/0       | top/cpu            |
| +program           |           | 14/14         | 48/48         | 29/29         | 1/1           | 4/4       | 0/0     | 0/0   | 0/0   | 0/0   | 0/0   | 0/0       | top/program        |
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+

Only 24 slices are used for the CPU!