Added outline of virtual machine
An outline of a virtual machine has been added to the Forth image, this
byte code interpreter will allow much more dense code to be made, and
will make the CPU easier to program, allowing a simple Forth interpreter
to be made. None of the assembly that makes up the virtual machine has
been tested, and it in fact is lacking a lot of machinery that would
allow it to work.

The idea is to create a byte code interpreter, each instruction being
8-bits in size, with an optional 8 or 16 bit operand. This, along with
the fact that the virtual machine is a dual stack based virtual machine
instead of an accumulator machine will allow software to be more
quickly developed. It will slow down execution by a fairly large
factor on an already slow machine.