Optional FIFO, working UART example code
* The FIFO for the UART can be turned on or off depending on a generic,
it does not increase the size of the project as much as I thought it
would. The UART module itself is far too large however and should be
replaced with something more svelte.
* The example code for the UART works, values can be pushed and popped
on the virtual machines stack which in turn feed into the UART, writing
the character 'H', this works in both the 'bit.c' simulator and the VHDL
simulator. It has not yet been tested in the hardware...
* The 'bit.c' simulator has been cleaned up a little, with better error
messages and the merging of two structures. Also, the load/store
instructions can now perform I/O like in the hardware.