104.619MHz: baud select, pc - 1, forth machinery
* The default baud rate of the system can be selected with by a generic,
this is useful for simulation as it allows a much higher baud rate to be
selected which means the processor spends less time waiting on the UART
to send characters. It also should be noted that because the system is
doing I/O and the I/O behaves slightly differently between the C
simulator and the VHDL that the output of the two systems has started to
diverge (it would be possible to bring them back together a little
more).
* A bug in the simulator has been found, an advanced program counter
value was retrieved by the C simulator iGET instruction.
* More of the Forth machinery necessary to make a working Forth
interpreter has been added to the 'bit.fth' image, although most of it
is not being used for now.